Start to create some real images.
I want to write some of the exercises that really render things as
integration tests that read from a file so that I can also view those
example files.

This commit adds the first of the integration tests. It doesn't quite
work yet.

It changes the main crate to be a lib rather than a bin (by removing the
`main.rs`).

This commit also updates the signature of the `canvas_to_ppm` function
by allowing it to take a mutable borrow. This was desirable to match
the interface of the `read_to_string` function.