Refactor `Tuple` struct into `Point` and `Vector`
These were the same type based on the direction of the book. I found
that there were a few places where I wanted to be able to lean on the
type system.

There is still some things I don't love about this:
- It's sorta weird that they both have `w` values.
- There are some places where I copy pasted implementations that could
  be extracted into private functions.

This is a very big commit and a lot of things changed.