Make the `Plane`'s and `Sphere`'s transformation matrix private.
This means that I had to use the "accessors" functions everywhere I was
getting or setting the transformation matrix.

I also made the names of the accessors more verbose. It's fine for the
field on the struct to be called `transform`, but since the trait is the
interface I might as well have a more clear name for what this attribute
of the shapes are.