Extract the writing of a file into a function.
The end to end tests work by writing the canvas to a file and then
reading the results into a string. We then assert if the result is the
exact same as before. I am not wild about this because it's possible
that some future feature will introduce some amount of non-determinism
and I mostly care that the renders don't changed significantly.