Add a `canvas_to_image` function
This function implemented on the Canvas struct returns an rgb image from the
`image` crate. This crate should allow writing different formats other
than the ppm format.

My plan is to remove the ppms and replace them with a different image
format (maybe png?) and change the tests to write fewer things to disk.
My theory is that this will make the tests run faster. Right now it
takes 18 seconds to run them all.