Add a builder interface to the World.
Worlds were being created via the modification of the `World` struct
directly. This meant that functions were aware of the implementation
details of the `World` struct would make the "multiple light sources"
TODO more difficult.

This commit introduces a new `WorldBuilder` struct which is intended to
contrain the adding of lights and the adding of shapes into an
interface.