Add an image gallery of what this project can render.
I had a number of images used as test fixtures that are worth showing
off, but were not easy to find.

This commit introduces a gallery section to the README. This commit also
introduces a CI check which should verify that all fixtures are included
in the gallery.

One follow up that I would like to do is have links to higher resolution
versions.